requirements.txt
faster-whisper
edge-tts
pydub
deep-translator
yt-dlp
demucs
soundfile
requests
google-genai
